Button Position changed and create common function
