+++ apps/front-end/public/locales/en/translation.json
@@ -254,6 +254,7 @@
   "DUE_TO_FAIL": "Failed",
   "DUE_TO_MARRIAGES": "Marriage",
   "DUPLICATE": "Duplicate",
+  "DUPLICATE_LIST": "Duplicate List",
   "E_AADHAAR_CARD_VERIFICATION": "E-Aadhaar card verification",
   "ECONOMIC_DIFFICULTIES": "Economic Difficulties",
   "EDIT_AADHAR_NUMBER": "Edit Adhar number",
