+++ apps/front-end/public/locales/hi/translation.json
@@ -252,6 +252,7 @@
   "DROPOUT_REASONS_PERSONAL_REASONS": "व्यक्तिगत कारण",
   "DROPOUTS": "ड्रॉपआउट",
   "DUPLICATE": "डुप्लिकेट",
+  "DUPLICATE_LIST": "डुप्लिकेट सूची",
   "DUE_TO_FAIL": "फेल होने के कारण",
   "DUE_TO_MARRIAGES": "शादी हो जाने के कारण",
   "E_AADHAAR_CARD_VERIFICATION": "ई-आधार कार्ड वेरिफिकेशन",
