+++ lib/common-lib/src/components/helper.js
@@ -649,3 +649,41 @@ export const setQueryParameters = (data) => {
     window.location.origin + window.location.pathname + jsonToQueryString(data)
   window.history.pushState('', '', pageUrl)
 }
+
+export const tableCustomStyles = {
+  rows: {
+    style: {
+      minHeight: '72px' // override the row height
+    },
+    style: {
+      minHeight: '72px' // override the row height
+    }
+  },
+  headCells: {
+    style: {
+      background: '#E0E0E0',
+      color: '#616161',
+      size: '16px'
+    },
+    style: {
+      background: '#E0E0E0',
+      color: '#616161',
+      size: '16px',
+      justifyContent: 'center'
+    }
+  },
+  cells: {
+    style: {
+      color: '#616161',
+      size: '19px'
+    },
+    style: {
+      color: '#616161',
+      size: '19px',
+      justifyContent: 'center'
+    }
+  }
+}
+
+
+
